refactor: Выделить селекторы в отдельный файл selectors.js
