+++ src/services/slices/burgerSlice.js
@@ -1,5 +1,4 @@
-import { createSelector, createSlice, nanoid } from '@reduxjs/toolkit';
-import { selectIngredientsEntities } from './ingredientsSlice';
+import { createSlice, nanoid } from '@reduxjs/toolkit';
 
 const initialState = {
   bun: null,
@@ -38,18 +37,6 @@ const burgerSlice = createSlice({
     moveIngredient,
   },
 });
-
-export const selectBurgerBun = (state) => state.burger.bun;
-export const selectBurgerIngredients = (state) => state.burger.ingredients;
-
-export const selectTotalPrice = createSelector(
-  [selectBurgerBun, selectBurgerIngredients, selectIngredientsEntities],
-  (bunId, ingredients, ingredientsEntities) =>
-    ingredients
-      .map(({ id }) => ingredientsEntities[id].price)
-      .reduce((total, price) => total + price, 0) +
-    (bunId ? ingredientsEntities[bunId].price * 2 : 0)
-);
 
 export const { actions } = burgerSlice;
 export default burgerSlice.reducer;
